% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{canopy_clst_plot}
\alias{canopy_clst_plot}
\title{Plotting the Canopy identified clusters}
\usage{
canopy_clst_plot(tinda_object, ...)
}
\arguments{
\item{tinda_object}{Object returned by TiNDA function}
}
\description{
Using ggplot2, plots VAF from tumor and control, and colors the based on canopy cluster
}
\examples{
 
tinda <- TiNDA(coverage_object, sample_name = "sample_3", data_type = "WGS")
canopy_clst_plot(tinda)

}
