% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{generate_depth}
\alias{generate_depth}
\title{Helper function to generate variant's depths}
\usage{
generate_depth(number_variants, avg_depth, sd = 5)
}
\arguments{
\item{number_variants}{Number of variants to simulate}

\item{avg_depth}{Average depth/coverage for the variants}

\item{sd}{Standard deviation of the coverage}
}
\description{
Helper function to generate variant's depths
}
