% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hg19_length}
\alias{hg19_length}
\title{Chromosome length}
\format{A data frame with chr name and its length
\itemize{
  \item CHR - chromosome name
  \item Length - Length of the chromosome
}}
\usage{
data(hg19_length)
}
\description{
Length of chromosomes for the linear TiNDA plots, default hg19. For used defined genomes, use the format below.
}
