% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Simulate_test_data}
\alias{Simulate_test_data}
\alias{generate_test_data}
\title{Data format for TiNDA input}
\format{
A data frame of rare germline variants
\itemize{
  \item CHR - chromosome name
  \item POS - variant position
  \item Control_ALT_DP - Read depth of the variant's alternate allele in the control sample
  \item Control_DP - Total read depth of the variant in the control sample
  \item Tumor_ALT_DP - Read depth of the variant's alternate allele in the tumor sample
  \item Tumor_DP - Total read_depth of the variant in the tumor sample
}
\tabular{rrrrrr}{
CHR \tab POS \tab Control_ALT_DP \tab Control_DP \tab Tumor_ALT_DP \tab Tumor_DP\cr
1 \tab 1039001 \tab 20 \tab 40 \tab 23 \tab 46\cr
1 \tab 2123023 \tab 12 \tab 32 \tab 14 \tab 23\cr
1 \tab 3343543 \tab 23 \tab 56 \tab 34 \tab 67
}
}
\usage{
generate_test_data(chr_length_table, num_variants_per_chr = 1000, ...)
}
\arguments{
\item{chr_length_table}{A data frame containing chromosome names and its length}

\item{...}{ellipsis

Simulate the test data generation for all the chromosomes}
}
\description{
The input format is VCF like data frame object containing information 
from both tumor and normal samples. It is recommended to filter out 
common SNPs using gnomAD or other population databases and 
only use the rare germline variants.
}
\details{
For user-defined input data, please use the same column names
as described below.
}
