% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tinda_clst_plot}
\alias{tinda_clst_plot}
\title{Plotting the  TiNDA indentified TiN clusters}
\usage{
tinda_clst_plot(tinda_object, ...)
}
\arguments{
\item{tinda_object}{Object returned by TiNDA function}
}
\description{
Plotting the somatic resuced and germline clusters as classified by TiNDA
}
\examples{
tinda <- TiNDA(coverage_object, sample_name = "sample_3", data_type = "WGS")
tinda_clst_plot(tinda)

}
