% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tinda_linear_plot}
\alias{tinda_linear_plot}
\title{Plotting linear genome}
\usage{
tinda_linear_plot(tinda_object, sample_af = "Tumor_AF",
  chr.length = NULL, colorCol = "TiN_Class", ...)
}
\arguments{
\item{tinda_object}{Object returned by TiNDA function}

\item{sample_af}{sample AF column in the TiNDA data, Default: 'Tumor_AF'}

\item{chr.length}{Data frame object of length of the chrs, if null hg19 chr length is used, Default: NULL'}

\item{colorCol}{Name of column with the final TiNDA classification, categorical variable, Default: 'TiN_Class'}
}
\description{
Function to plot choromosomes linerlly from TiNDA object
}
\examples{
tinda <- TiNDA(coverage_object, sample_name = "sample_3", data_type = "WGS")
plotGenome_ggplot(tinda)
plotGenome_ggplot(tinda, sample_af = 'Control_AF')

}
