% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tinda_summary_plot}
\alias{tinda_summary_plot}
\title{TiNDA summary plot}
\usage{
tinda_summary_plot(tinda_object, ...)
}
\arguments{
\item{tinda_object}{Object returned by TiNDA function}
}
\description{
A summary plot combines differnt TiNDA plots and TiNDA results
}
\examples{

}
