% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_data}
\alias{write_data}
\title{Write TiNDA data}
\usage{
write_data(tinda_object, file_name)
}
\arguments{
\item{tinda_object}{TiNDA object produced by the main TiNDA function}

\item{file_name}{Name of the file}
}
\description{
Function to write TiNDA data to a tsv file with TiN_Class information
}
