% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{canopy_clst_plot}
\alias{canopy_clst_plot}
\title{Plotting the Canopy identified clusters}
\usage{
canopy_clst_plot(tinda_object, ...)
}
\arguments{
\item{tinda_object}{Object returned by TiNDA function}

\item{...}{ellipsis}
}
\description{
Using ggplot2, plots VAF from tumor and control, and colors the based on canopy cluster
}
\examples{
 
data(hg19_length)
vcf_like_df = TiNDA::generate_test_data(hg19_length)
tinda_test_object <- TiNDA(vcf_like_df, sample_name = "sample_3", data_type = "WGS")
canopy_clst_plot(tinda_test_object)

}
