% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tinda_summary_plot}
\alias{tinda_summary_plot}
\title{TiNDA summary plot}
\usage{
tinda_summary_plot(tinda_object, ...)
}
\arguments{
\item{tinda_object}{Object returned by TiNDA function}

\item{...}{ellipsis}
}
\description{
A summary plot combines different TiNDA plots and TiNDA results
}
\examples{
data(hg19_length)
vcf_like_df = TiNDA::generate_test_data(hg19_length)
tinda_test_object <- TiNDA(vcf_like_df, sample_name = "sample_3", data_type = "WGS")
tinda_summary_plot(tinda_test_object)
 
}
